display number of reminders needs in side menu
