+++ app/helpers/badges_helper.rb
@@ -5,4 +5,8 @@ module BadgesHelper
     tag.div(badge.title, class: 'ui basic label badge',
                 style: "border: 1px solid #{badge.color}; color: #{badge.color}")
   end
+
+  def simple_label(content, color = '')
+    tag.div(content, class: "ui label #{color}")
+  end
 end
