+++ app/views/reminders/experts/index.haml
@@ -1,7 +1,7 @@
 - meta title: t('.title')
 
 - content_for :header, render('reminders/header', title: t('.title'))
-- content_for :menu, render('reminders/menu')
+- content_for :menu, render('reminders/menu', count_needs: @count_needs)
 
 = form_with url: reminders_experts_path, method: :get, skip_enforcing_utf8: true, class: 'ui large form' do
   - options = { include_blank: t('all_choices'), class: 'ui search selection dropdown', onchange: 'this.form.submit();' }
