+++ app/views/reminders/needs/index.html.haml
@@ -1,5 +1,5 @@
 - content_for :header, render('reminders/header', title: t('.title', status: @status))
-- content_for :menu, render('reminders/menu')
+- content_for :menu, render('reminders/menu', count_needs: @count_needs)
 
 = form_with url: request.path, method: :get, skip_enforcing_utf8: true, class: 'ui large form' do
   - options = { include_blank: t('all_choices'), class: 'ui search selection dropdown', onchange: 'this.form.submit();' }
