+++ config/locales/views.fr.yml
@@ -554,9 +554,9 @@ fr:
       index:
         title: Besoins %{status}
       menu:
-        abandoned_without_taking_care_html: Abandonnés <i>(+30 jours)</i>
-        reminder_in_progress_html: En cours <i>(relancés)</i>
-        reminder_quo_not_taken_html: À traiter <i>(+10 jours)</i>
+        abandoned_without_taking_care_html: "%{label} Abandonnés <i>(+30 jours)</i>"
+        reminder_in_progress_html: "%{label} Relancés"
+        reminder_quo_not_taken_html: "%{label} À traiter <i>(+10 jours)</i>"
   solicitations:
     index:
       clear_search: Réinitialiser la recherche
